Normalize empty lines between entries
Having the entries separated by 2 empty lines makes it easier for rst
newcomers to find their way through the file.  It also fixes the entry
for Nectar API, which previously was not recognized as a heading.

To normalize the empty lines between the entries:

~~~text
replace: ([^\n])\n*(`.*`__\n~~~~)
with: $1\n\n\n$2
~~~

To normalize the length of the lines below the headings:

~~~text
replace: ^~~~~~+
with: ~~~~
~~~

The point is that the lines were quite inconsistent because for one-time
editors it's hard to see that the line is supposed to be as long as the
heading directly above it.

Having the same number of tildes makes it easier to edit the document
using copy-and-paste and still keep a consistent style.