Use images directly
This makes it easier to add an entry to the file since there is only a
single place to modify instead of the current 2 places.

The cleanup has been done by the following Python program:

~~~python
import re

with open("README.rst") as f:
    image = {}
    for line in f:
        m = re.search(r"^\.\. (\|.*\|) image:: (.*)$", line.rstrip())
        if m:
            image[m[1]] = m[2]

lines = []
with open("README.rst") as f:
    for line in f:
        line = line.rstrip()
        if line in image:
            lines.append(".. image:: " + image[line])
            del image[line]
            continue
        m = re.search(r"^\.\. (\|.*\|) image:: (.*)$", line.rstrip())
        if m and m[1] not in image:
            continue
        lines.append(line)

with open("README.rst", "w") as f:
    for line in lines:
        f.write(line + "\n")

print(repr(image))
~~~

It left some entries that have to be cleaned up manually since they were
wrong.