Normalize empty lines between entries
Having the entries separated by 2 empty lines makes it easier for rst
newcomers to find their way through the file.  It also fixes the entry
for Nectar API, which previously was not recognized as a heading.

To normalize the empty lines between the entries:

~~~text
replace: ([^\n])\n*(`.*`__\n~~~~)
with: $1\n\n\n$2
~~~